[Enhance] Move TransactionInfo and TransactionContent to global component folder.
