+++ src/components/TransactionInfo.tsx
@@ -1,8 +1,14 @@
 import { Box, Text } from "@chakra-ui/react";
-import React, { ReactNode, useCallback, useRef, useState } from "react";
+import React, {
+  ReactNode,
+  useCallback,
+  useEffect,
+  useRef,
+  useState,
+} from "react";
 import { FormattedMessage } from "react-intl";
 import measureTextWidth from "src/utils/measureTextWidth";
-import messages from "./messages";
+import messages from "../features/EVM/Transaction/components/messages";
 
 interface TransactionDetail {
   hasEnoughBalance: boolean;
@@ -21,17 +27,20 @@ const TransactionInfo = ({
   children: ReactNode;
   transactionDetail?: TransactionDetail;
 }) => {
-  const [showTokenAmount, setShowTokenAmount] = useState(false);
   const {
     usdValue = "",
     tokenName = "",
     tokenAmount = "",
   } = transactionDetail || {};
+  const [showTokenAmount, setShowTokenAmount] = useState(false);
+  const [urlFontSize, setUrlFontSize] = useState(0);
+  const [tokenFontSize, setTokenFontSize] = useState(0);
+
   const tokenAmountRef = useRef<HTMLDivElement>(null);
   const rootRef = useRef<HTMLDivElement>(null);
 
   const getFitFontSize = useCallback((text: string, maxFontSize = 36) => {
-    if (!rootRef.current) return;
+    if (!rootRef.current) return 0;
 
     const textWidth = measureTextWidth(text, `${maxFontSize}px 'Work Sans'`);
     const rootWidth = rootRef.current?.clientWidth || 0;
@@ -42,13 +51,23 @@ const TransactionInfo = ({
       maxFontSize
     ).toFixed(2);
 
+    const MIN_FONT_SIZE = 16;
     const scaledFontSize = Math.min(
-      Math.max(16, +elementScaleValue),
+      Math.max(MIN_FONT_SIZE, +elementScaleValue),
       maxFontSize
     );
 
     return scaledFontSize;
   }, []);
+
+  useEffect(() => {
+    setUrlFontSize(getFitFontSize(host, 16));
+    setTokenFontSize(getFitFontSize(tokenAmount, 28));
+
+    if (!transactionDetail) return;
+
+    setShowTokenAmount(true);
+  }, [transactionDetail, getFitFontSize, host, setTokenFontSize, tokenAmount]);
 
   return (
     <Box
@@ -67,7 +86,7 @@ const TransactionInfo = ({
         px="space.m"
         py="space.3xs"
         borderRadius="32px"
-        fontSize="size.subheading.2"
+        fontSize={`${urlFontSize}px`}
         fontWeight="weight.m"
         mb="space.2xs"
       >
@@ -79,12 +98,14 @@ const TransactionInfo = ({
             ref={tokenAmountRef}
             mb="space.4xs"
             fontWeight="weight.l"
-            fontSize={getFitFontSize(tokenAmount, 28)}
+            fontSize={`${tokenFontSize}px`}
             lineHeight="line.height.heading.2"
           >
-            {`${tokenAmount} ${tokenName}`}
+            {showTokenAmount && `${tokenAmount} ${tokenName}`}
           </Box>
-          <Box color="font.secondary">≈ {usdValue} USD</Box>
+          {showTokenAmount && (
+            <Box color="font.secondary">≈ {usdValue} USD</Box>
+          )}
         </>
       )}
     </Box>
